+++ CHANGELOG.md
@@ -127,7 +127,7 @@ Percentage | Module
   * [Kernel] Warn on unsupported nested comparisons such as `x < y < z`
   * [Kernel] Warn if redefining documentation across clauses of the same definition
   * [Kernel] Warn on unnecessary quotes around atoms, keywords and calls
-  * [Macro] Add `Macro.special_form?/2` and `Macro.operator?/2` that returns true if the given name/arity is a special form or operator respectively
+  * [Macro] Add `Macro.special_form?/2` and `Macro.operator?/2` that returns `true` if the given name/arity is a special form or operator respectively
   * [Macro.Env] Add `Macro.Env.vars/1` and `Macro.Env.has_var?/2` that gives access to environment data without accessing private fields
   * [Regex] Include endianness in the regex version. This allows regexes to be recompiled when an archive is installed in a system with a different endianness
   * [Registry] Add `Registry.count/1` and `Registry.count_match/4`